ws/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.37.10
botocore==1.37.10
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
jmespath==1.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
watchtower==3.4.0
websockets==15.0.1
